Adjusting speed works. Slow is awesome, fast is very crackly, not sure if just emulator. Scales 0.xxx to 2.0

+++ BeatMatrix/src/com/joshuac/beatmatrix/ButtonMatrix.java
@@ -97,7 +97,7 @@ public class ButtonMatrix extends Activity implements ChooseFileDialog.OnChooseF
 		showSongEditDialog();
     }
 	
-	public void onEditInfoSelected(double start_time, double end_time, float volume, double speed) {
+	public void onEditInfoSelected(double start_time, double end_time, float volume, float speed) {
 		
 		//chooseButton.setImageDrawable(getResources().getDrawable(R.drawable.playlist_off)); 
 		//Consider deleting
